# Improve performance of tests